Don't copy DirectoryOpus.CFG because that file works only on 32-bit little endian. Various compiler warnings fixed.
DirectoryOpus starts on x86_64. Short tests of some
file operations work good. It sometimes crashes on
exit, so there are likely some cases where memory
is allocated with wrong size.


git-svn-id: https://svn.aros.org/svn/aros/trunk@54894 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/dopus/DOpus_Disk/diskop.c
@@ -408,7 +408,7 @@ char *alike;
       
     while ((devnode = (struct DeviceNode *)NextDosEntry((struct DosList *)devnode,LDF_DEVICES))) {
         if (/*devnode->dn_Type==DLT_DEVICE && devnode->dn_Task &&*/
-            devnode->dn_Startup>512) {
+            (long)devnode->dn_Startup>512) {
             if (!alikenode || like_devices(devnode,alikenode)) 
                 ++count;
         }
@@ -422,7 +422,7 @@ char *alike;
         while ((devnode = (struct DeviceNode *)NextDosEntry((struct DosList *)devnode,LDF_DEVICES))) {
 //        while (devnode) {
             if (/*devnode->dn_Type==DLT_DEVICE && devnode->dn_Task &&*/
-                devnode->dn_Startup>512) {
+                (long)devnode->dn_Startup>512) {
                 if (!alikenode || like_devices(devnode,alikenode)) {
                     BtoCStr((BPTR)devnode->dn_Name,devname,32);
                     strcat(devname,":");
